requirements.txt
websocket-client
kafka-python
pyspark
duckdb
pandas 